Refactor leaflet map JS components
