+++ src/scenes/papers/components/panel/map/MapInfoBox.js
