+++ src/scenes/papers/components/panel/map/MapSelectPaper.js
